Work with live-reload
This updates steal-qunit to work with live-reload. Whenever a QUnit test
module is reloaded its tests will be reran; this already works. This
change just makes it so that steal-qunit will update the UI to remove
extra elements that are added and updates the qunit-banner to show the
tests as either passing or failing.

Closes #7